requirements.txt
certifi==2017.11.5
chardet==3.0.4
future==0.16.0
idna==2.6
oauthlib==2.0.6
pymongo==3.6.0
python-twitter==3.3
requests==2.18.4
requests-oauthlib==0.8.0
urllib3==1.22
